alioth: overlays: Move overlays to product While previous Android versions respected priorities regardless of partition, Android 14 prioritizes partitions.
Overlay precedence:
When multiple overlays override the same resources, the order of the
overlays is important. An overlay has greater precedence than overlays
with configurations preceding its own configuration. The precedence
order of overlays in different partitions (from least to greatest
precedence) is as follows:

system
vendor
odm
oem
product
system_ext

This is a shame for GSI compatibility, but we can't override
otherwise.

Change-Id: I688cbf2cc15725189c70c650092f76637f8d90fc